+++ backend/pi-planning-backend/Migrations/20260215065624_InitialCreate.Designer.cs
@@ -12,7 +12,7 @@ using PiPlanningBackend.Data;
 namespace PiPlanningBackend.Migrations
 {
     [DbContext(typeof(AppDbContext))]
-    [Migration("20251026200106_InitialCreate")]
+    [Migration("20260215065624_InitialCreate")]
     partial class InitialCreate
     {
         /// <inheritdoc />
@@ -192,11 +192,11 @@ namespace PiPlanningBackend.Migrations
 
                     NpgsqlPropertyBuilderExtensions.UseIdentityByDefaultColumn(b.Property<int>("Id"));
 
-                    b.Property<double>("CapacityDev")
-                        .HasColumnType("float");
+                    b.Property<int>("CapacityDev")
+                        .HasColumnType("integer");
 
-                    b.Property<double>("CapacityTest")
-                        .HasColumnType("float");
+                    b.Property<int>("CapacityTest")
+                        .HasColumnType("integer");
 
                     b.Property<int>("SprintId")
                         .HasColumnType("integer");
@@ -290,7 +290,7 @@ namespace PiPlanningBackend.Migrations
             modelBuilder.Entity("PiPlanningBackend.Models.TeamMember", b =>
                 {
                     b.HasOne("PiPlanningBackend.Models.Board", "Board")
-                        .WithMany()
+                        .WithMany("TeamMembers")
                         .HasForeignKey("BoardId")
                         .OnDelete(DeleteBehavior.Cascade)
                         .IsRequired();
@@ -341,6 +341,8 @@ namespace PiPlanningBackend.Migrations
                     b.Navigation("Features");
 
                     b.Navigation("Sprints");
+
+                    b.Navigation("TeamMembers");
                 });
 
             modelBuilder.Entity("PiPlanningBackend.Models.Feature", b =>